Holy

+++ Schule/src/blatt26/aufgabe01/Punkt.java
@@ -1,10 +1,10 @@
 package blatt26.aufgabe01;
 
 public class Punkt {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Punkt(int x, int y) {
+    public Punkt(double x, double y) {
         this.x = x;
         this.y = y;
     }
